// backend/src/containers/enums/container-type.enum.ts
// backend\src\containers\enums\container-type.enum.ts
export enum ContainerType {
    GENERAL = 'GENERAL',
    PAPER = 'PAPER',
    PLASTIC = 'PLASTIC',
    GLASS = 'GLASS',
    ORGANIC = 'ORGANIC',
    METAL = 'METAL',
    ELECTRONICS = 'ELECTRONICS',
    BATTERIES = 'BATTERIES',
    CLOTHING = 'CLOTHING',
    OIL = 'OIL',
    OTHERS = 'OTHERS',
}